Forbid deletion of godowns which are already referred in transactions

Closes #954

See merge request gnukhata/gkcore!739

+++ src/gkcore/views/api_godown.py
@@ -429,6 +429,15 @@ class api_godown(object):
             try:
                 self.con = eng.connect()
                 dataset = self.request.json_body
+                is_godown_used = self.con.execute(
+                    select([stock.c.goid]).where(stock.c.goid == dataset["goid"])
+                ).rowcount
+                if is_godown_used:
+                    return {
+                        "gkstatus": enumdict["ActionDisallowed"],
+                        "error": "Cannot delete godowns already referred in transactions",
+                    }
+
                 result = self.con.execute(
                     godown.delete().where(godown.c.goid == dataset["goid"])
                 )
